+++ googlesource/acceptance_test.go
@@ -25,86 +25,16 @@ import (
 	"go.uber.org/goleak"
 )
 
-// type destination struct {
-// 	sdk.UnimplementedDestination
-// }
-
-// func NewDestination() sdk.Destination {
-// 	return destination{}
-// }
-
-// func (d destination) Configure(context.Context, map[string]string) error        { return nil }
-// func (d destination) Open(context.Context) error                                { return nil }
-// func (d destination) WriteAsync(context.Context, sdk.Record, sdk.AckFunc) error { return nil }
-// func (d destination) Write(context.Context, sdk.Record) error                   { return nil }
-// func (d destination) Flush(context.Context) error                               { return nil }
-// func (d destination) Teardown(context.Context) error                            { return nil }
-
-// func TestAcceptance(t *testing.T) {
-
-// 	cfg := map[string]string{
-// 		Servers:           "localhost:9092",
-// 		ReadFromBeginning: "true",
-// 	}
-
-// 	sdk.AcceptanceTest(t, driver{
-// 		ConfigurableAcceptanceTestDriver: sdk.ConfigurableAcceptanceTestDriver{
-// 			Config: sdk.ConfigurableAcceptanceTestDriverConfig{
-// 				Connector: sdk.Connector{
-// 					NewSpecification: Specification,
-// 					NewSource:        NewSource,
-// 					NewDestination:   NewDestination,
-// 				},
-// 				SourceConfig:      cfg,
-// 				DestinationConfig: cfg,
-
-// 				BeforeTest: func(t *testing.T) {
-// 					cfg[Topic] = "TestAcceptance-" + uuid.NewString()
-// 				},
-// 				AfterTest: func(t *testing.T) {
-// 				},
-// 				GoleakOptions: []goleak.Option{
-// 					// kafka.DefaultTransport starts some goroutines: https://github.com/segmentio/kafka-go/issues/599
-// 					goleak.IgnoreTopFunction("github.com/segmentio/kafka-go.(*connPool).discover"),
-// 					goleak.IgnoreTopFunction("github.com/segmentio/kafka-go.(*conn).run"),
-// 				},
-// 			},
-// 		},
-// 	})
-// 	// cfg := map[string]string{
-// 	// 	googlebigquery.ConfigServiceAccount:     serviceAccount,
-// 	// 	googlebigquery.ConfigProjectID:          projectID,
-// 	// 	googlebigquery.ConfigDatasetID:          datasetID,
-// 	// 	googlebigquery.ConfigTableID:            tableIDTimeStamp,
-// 	// 	googlebigquery.ConfigLocation:           location,
-// 	// 	googlebigquery.ConfigIncrementalColName: fmt.Sprintf("%s:updatedat", tableIDTimeStamp),
-// 	// 	googlebigquery.ConfigPrimaryKeyColName:  fmt.Sprintf("%s:age", tableIDTimeStamp),
-// 	// }
-
-// 	// sdk.AcceptanceTest(t, AcceptanceTestDriver{
-// 	// 	Config: AcceptanceSourceTestDriverConfig{
-// 	// 		Connector: sdk.Connector{ // Note that this variable should rather be created globally in `connector.go`
-// 	// 			NewSpecification: googlebigquery.Specification,
-// 	// 			NewSource:        NewSource,
-// 	// 			// NewDestination:   NewDestination,
-// 	// 		},
-// 	// 		SourceConfig: cfg,
-
-// 	// 		BeforeTest: func(t *testing.T) {
-// 	// 			DataSetup()
-// 	// 			time.Sleep(10 * time.Second)
-// 	// 		},
-// 	// 	},
-// 	// })
-// }
-
 func TestAcceptance(t *testing.T) {
 	cfg := map[string]string{
-		googlebigquery.ConfigServiceAccount: serviceAccount,
-		googlebigquery.ConfigProjectID:      projectID,
-		googlebigquery.ConfigDatasetID:      datasetID,
-		googlebigquery.ConfigTableID:        "table_acceptance",
-		googlebigquery.ConfigLocation:       location,
+		googlebigquery.ConfigServiceAccount:     serviceAccount,
+		googlebigquery.ConfigProjectID:          projectID,
+		googlebigquery.ConfigDatasetID:          datasetID,
+		googlebigquery.ConfigTableID:            "table_acceptance",
+		googlebigquery.ConfigLocation:           location,
+		googlebigquery.ConfigPrimaryKeyColName:  "table_acceptance:name", //"table_acceptance"
+		googlebigquery.ConfigIncrementalColName: "table_acceptance:name",
+		googlebigquery.ConfigPollingTime:        "1ms",
 	}
 
 	sdk.AcceptanceTest(t, AcceptanceTestDriver{
@@ -113,20 +43,23 @@ func TestAcceptance(t *testing.T) {
 				NewSpecification: googlebigquery.Specification,
 				NewSource:        NewSource,
 			},
-			SourceConfig:  cfg,
+			SourceConfig: cfg,
 			GoleakOptions: []goleak.Option{
-				// kafka.DefaultTransport starts some goroutines: https://github.com/segmentio/kafka-go/issues/599
-				// goleak.IgnoreTopFunction("github.com/segmentio/kafka-go.(*connPool).discover"),
-				// goleak.IgnoreTopFunction("github.com/segmentio/kafka-go.(*conn).run"),
+				goleak.IgnoreTopFunction("go.opencensus.io/stats/view.(*worker).start"), //indirect leak from dependency go.opencensus.io
+				goleak.IgnoreTopFunction("internal/poll.runtime_pollWait"),
+			},
+			AfterTest: func(t *testing.T) {
+				err := cleanupDataset([]string{cfg[googlebigquery.ConfigTableID]})
+				if err != nil {
+					fmt.Println("Error found")
+				}
 			},
 		},
 	},
 	)
 }
 
-// AcceptanceTestDriver is the default implementation of
-// AcceptanceTestDriver. It provides a convenient way of configuring the driver
-// without the need of implementing a custom driver from scratch.
+// AcceptanceTestDriver implements sdk.AcceptanceTestDriver
 type AcceptanceTestDriver struct {
 	Config AcceptanceSourceTestDriverConfig
 }
@@ -137,8 +70,7 @@ type AcceptanceSourceTestDriverConfig struct {
 	// Connector is the connector to be tested.
 	Connector sdk.Connector
 
-	// SourceConfig should be a valid config for a source connector, reading
-	// from the same location as the destination will write to.
+	// SourceConfig config for source
 	SourceConfig map[string]string
 
 	// BeforeTest is executed before each acceptance test.
@@ -200,46 +132,13 @@ func (d AcceptanceTestDriver) GoleakOptions(_ *testing.T) []goleak.Option {
 	return d.Config.GoleakOptions
 }
 
-// WriteToSource by default opens the destination and writes records to the
-// destination. It is expected that the destination is writing to the same
-// location the source is reading from. If the connector does not implement a
-// destination the function will fail the test.
+// WriteToSource writes data for source to pull data from
 func (d AcceptanceTestDriver) WriteToSource(t *testing.T, records []sdk.Record) []sdk.Record {
-	// if d.Connector().NewDestination == nil {
-	// 	t.Fatal("connector is missing the field NewDestination, either implement the destination or overwrite the driver method Write")
-	// }
 	var err error
 	is := is.New(t)
-	// ctx, cancel := context.WithCancel(context.Background())
-	// defer cancel()
 	config := d.SourceConfig(t)
-	fmt.Println("**config", config)
-	fmt.Println("*****records :", len(records))
 	records, err = writeToSource(config, records)
 	is.NoErr(err)
-	// d.WriteToSource()
-
-	// // writing something to the destination should result in the same record
-	// // being produced by the source
-	// dest := d.Connector().NewDestination()
-	// err := dest.Configure(ctx, d.DestinationConfig(t))
-	// is.NoErr(err)
-
-	// err = dest.Open(ctx)
-	// is.NoErr(err)
-
-	// defer func() {
-	// 	cancel() // cancel context to simulate stop
-	// 	err = dest.Teardown(context.Background())
-	// 	is.NoErr(err)
-	// }()
-
-	// // try to write using WriteAsync and fallback to Write if it's not supported
-	// err := d.writeAsync(ctx, dest, records)
-	// if errors.Is(err, sdk.ErrUnimplemented) {
-	// err := d.write(ctx, records)
-	// // }
-	// is.NoErr(err)
 
 	return records
 }
@@ -251,23 +150,3 @@ func writeToSource(config map[string]string, records []sdk.Record) (result []sdk
 func (d AcceptanceTestDriver) ReadFromDestination(*testing.T, []sdk.Record) []sdk.Record {
 	return []sdk.Record{}
 }
-
-// // write writes records to destination using Destination.Write.
-// func (d AcceptanceTestDriver) write(ctx context.Context, records []sdk.Record) error {
-// 	for _, r := range records {
-// 		err := dest.Write(ctx, r)
-// 		if err != nil {
-// 			return err
-// 		}
-// 	}
-
-// 	// flush to make sure the records get written to the destination, but allow
-// 	// it to be unimplemented
-// 	err := dest.Flush(ctx)
-// 	if err != nil && !errors.Is(err, sdk.ErrUnimplemented) {
-// 		return err
-// 	}
-
-// 	// records were successfully written
-// 	return nil
-// }
